Moved files to seperates folders, fixed links
