+++ src/components/offers-card/offer-card.tsx
